Enemigo Que se muere

+++ Assets/_Main/Scripts/Gun/WeaponStats.cs
@@ -5,8 +5,8 @@ public class WeaponStats : ScriptableObject {
     [SerializeField] private float fireRate;
     public float FireRate => fireRate;
 
-    [SerializeField] private float weaponDamage;
-    public float WeaponDamage => weaponDamage;
+    [SerializeField] private int weaponDamage;
+    public int WeaponDamage => weaponDamage;
 
     [SerializeField] private Bullet bulletPrefab;
     public Bullet BulletPrefab => bulletPrefab;
